// backend/src/generated/models/TemplateSection.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TemplateSection = {
    /**
     * Unique identifier for the section
     */
    id: string;
    /**
     * Display name of the section
     */
    name: string;
    /**
     * Description of what should be included in this section
     */
    description: string;
    /**
     * Placeholder text for the section
     */
    placeholder?: string;
    /**
     * Whether this section is required
     */
    required?: boolean;
    /**
     * Display order of the section
     */
    order: number;
};